test(tracer): separate CoreTracerBasic and CoreTracer tests
